#515: Combined StorageMap integration tests given the max tx size change
