+++ compiler/src/test-integration/java/io/neow3j/compiler/StorageMapIntegrationTest.java
@@ -1,11 +1,13 @@
 package io.neow3j.compiler;
 
-import io.neow3j.types.ContractParameter;
 import io.neow3j.devpack.ByteString;
 import io.neow3j.devpack.Storage;
 import io.neow3j.devpack.StorageContext;
 import io.neow3j.devpack.StorageMap;
 import io.neow3j.protocol.core.response.InvocationResult;
+import io.neow3j.types.ContractParameter;
+import io.neow3j.types.Hash160;
+import io.neow3j.types.Hash256;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -16,6 +18,8 @@ import java.io.IOException;
 
 import static io.neow3j.types.ContractParameter.byteArray;
 import static io.neow3j.types.ContractParameter.byteArrayFromString;
+import static io.neow3j.types.ContractParameter.hash160;
+import static io.neow3j.types.ContractParameter.hash256;
 import static io.neow3j.types.ContractParameter.integer;
 import static io.neow3j.types.ContractParameter.string;
 import static io.neow3j.devpack.StringLiteralHelper.hexToBytes;
@@ -107,6 +111,24 @@ public class StorageMapIntegrationTest {
     }
 
     @Test
+    public void putByteArrayKeyHash160Value() throws IOException {
+        Hash160 v = ct.getDefaultAccount().getScriptHash();
+        ContractParameter key = byteArray("02");
+        ContractParameter value = hash160(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getAddress(), is(v.toAddress()));
+    }
+
+    @Test
+    public void putByteArrayKeyHash256Value() throws IOException {
+        Hash256 v = ct.getDeployTxHash();
+        ContractParameter key = byteArray("02");
+        ContractParameter value = hash256(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getByteArray(), is(v.toLittleEndianArray()));
+    }
+
+    @Test
     public void putByteStringKeyByteArrayValue() throws IOException {
         String v = "050607";
         ContractParameter key = byteArray("02");
@@ -143,6 +165,24 @@ public class StorageMapIntegrationTest {
     }
 
     @Test
+    public void putByteStringKeyHash160Value() throws IOException {
+        Hash160 v = ct.getDefaultAccount().getScriptHash();
+        ContractParameter key = byteArray("02");
+        ContractParameter value = hash160(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getAddress(), is(v.toAddress()));
+    }
+
+    @Test
+    public void putByteStringKeyHash256Value() throws IOException {
+        Hash256 v = ct.getDeployTxHash();
+        ContractParameter key = byteArray("02");
+        ContractParameter value = hash256(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getByteArray(), is(v.toLittleEndianArray()));
+    }
+
+    @Test
     public void putStringKeyByteArrayValue() throws IOException {
         String v = "050607";
         ContractParameter key = string("aa");
@@ -176,6 +216,24 @@ public class StorageMapIntegrationTest {
         ContractParameter value = string(v);
         InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
         assertThat(res.getStack().get(0).getString(), is(v));
+    }
+
+    @Test
+    public void putStringKeyHash160Value() throws IOException {
+        Hash160 v = ct.getDefaultAccount().getScriptHash();
+        ContractParameter key = string("aa");
+        ContractParameter value = hash160(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getAddress(), is(v.toAddress()));
+    }
+
+    @Test
+    public void putStringKeyHash256Value() throws IOException {
+        Hash256 v = ct.getDeployTxHash();
+        ContractParameter key = string("aa");
+        ContractParameter value = hash256(v);
+        InvocationResult res = ct.callInvokeFunction(testName, key, value).getInvocationResult();
+        assertThat(res.getStack().get(0).getByteArray(), is(v.toLittleEndianArray()));
     }
 
     @Test
@@ -247,6 +305,18 @@ public class StorageMapIntegrationTest {
             return Storage.get(ctx, prefix.concat(key));
         }
 
+        public static ByteString putByteArrayKeyHash160Value(byte[] key,
+                io.neow3j.devpack.Hash160 value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
+        public static ByteString putByteArrayKeyHash256Value(byte[] key,
+                io.neow3j.devpack.Hash256 value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
         public static ByteString putByteStringKeyByteArrayValue(ByteString key, byte[] value) {
             map.put(key, value);
             return Storage.get(ctx, prefix.concat(key));
@@ -267,6 +337,18 @@ public class StorageMapIntegrationTest {
             return Storage.get(ctx, prefix.concat(key));
         }
 
+        public static ByteString putByteStringKeyHash160Value(ByteString key,
+                io.neow3j.devpack.Hash160 value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
+        public static ByteString putByteStringKeyHash256Value(ByteString key,
+                io.neow3j.devpack.Hash256 value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
         public static ByteString putStringKeyByteArrayValue(String key, byte[] value) {
             map.put(key, value);
             return Storage.get(ctx, prefix.concat(key));
@@ -283,6 +365,18 @@ public class StorageMapIntegrationTest {
         }
 
         public static ByteString putStringKeyStringValue(String key, String value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
+        public static ByteString putStringKeyHash160Value(String key,
+                io.neow3j.devpack.Hash160 value) {
+            map.put(key, value);
+            return Storage.get(ctx, prefix.concat(key));
+        }
+
+        public static ByteString putStringKeyHash256Value(String key,
+                io.neow3j.devpack.Hash256 value) {
             map.put(key, value);
             return Storage.get(ctx, prefix.concat(key));
         }
